refactor(watch): extract useClaimProbe; share near-end constant from watchStallRecovery
